moves isolated AGSLicense+Description to caller file
